refactor : repository 패턴 도입
및 api call tryCatch 중복 로직을 공통 모듈로 분리하여 관리.(HomeViewModel)
