+++ app/src/main/java/com/echoist/linkedout/HomeRepository.kt
@@ -2,7 +2,10 @@ package com.echoist.linkedout
 
 import android.util.Log
 import com.echoist.linkedout.api.SupportApi
+import com.echoist.linkedout.api.UserApi
 import com.echoist.linkedout.data.NotificationResponse
+import com.echoist.linkedout.data.UserGraphSummaryResponse
+import com.echoist.linkedout.data.UserInfo
 import com.echoist.linkedout.page.myLog.Token
 import retrofit2.Response
 import javax.inject.Inject
@@ -10,6 +13,7 @@ import javax.inject.Inject
 
 class HomeRepository @Inject constructor(
     private val supportApi: SupportApi,
+    private val userApi: UserApi
 ) {
 
     suspend fun requestUserNotification(
@@ -18,6 +22,27 @@ class HomeRepository @Inject constructor(
     ) {
         apiCall(onSuccess = onSuccess, finally = finally)
         { supportApi.readUserNotification(Token.bearerAccessToken, Token.refreshToken) }
+    }
+
+    suspend fun requestFirstUserToExistUser(){
+        val isNotFirst = UserInfo(isFirst = false)
+        apiCall(
+            onSuccess = {
+                Log.d("첫유저 ->기존유저", "성공")
+            },
+            onError = { e ->
+                Log.e("첫유저 ->기존유저", "실패 ${e.message}")
+            }
+        ) { userApi.userUpdate(Token.bearerAccessToken, Token.refreshToken, isNotFirst) }
+    }
+
+    suspend fun requestUserGraphSummary(onSuccess: suspend (UserGraphSummaryResponse) -> Unit,) {
+        apiCall(
+            onSuccess = onSuccess ,
+            onError = { e ->
+                Log.e("유저 주간 링크드아웃 지수:", "${e.message}")
+            }
+        ) { userApi.requestUserGraphSummary(Token.bearerAccessToken, Token.refreshToken) }
     }
 
 }
